edemco-backend-python/microservice_growatt-generation/requirements.txt
selenium==4.19.0
xlrd==2.0.1
pandas==2.2.2
Flask==3.0.3
Flask-Cors==4.0.1
python-dateutil==2.9.0.post0